requirements.txt
openai
python-dotenv
sklearn